# app/assets/swagger/v1.yaml
openapi: 3.0.0
info:
  title: OIDC VPN Manager Frontend API
  version: "1.0.0"
  description: |
    REST API for OIDC VPN Manager device and server authentication.
    
    This API provides PSK-authenticated endpoints for:
    - Server configuration bundle retrieval for automated deployment (admin service only)
    
    ## Service Separation
    
    The frontend service supports three deployment patterns:
    - **Combined Service**: All endpoints available on single service
    - **User Service**: Web UI available for user certificate generation via OIDC, admin API endpoints return 403
    - **Admin Service**: Admin API endpoints available, user routes redirect to user service
    
    ## Authentication
    
    All API endpoints require authentication using Pre-Shared Keys (PSK) via Bearer token:
    ```
    Authorization: Bearer your-psk-here
    ```
    
    PSKs are generated through the administrative web interface and must be distributed securely to authorized devices and servers. Each PSK is associated with a specific description/hostname for audit purposes.
    
    ## Security
    
    - All endpoints require HTTPS in production
    - PSKs should be stored securely and rotated regularly
    - Each PSK is associated with a specific description for audit purposes
    - Certificate generation includes client IP tracking for geolocation audit trails
  contact:
    email: jon+openvpnmanagersecurity@sprig.gs
  license:
    name: AGPL v3
...
servers:
...
paths:
  /openvpn-api/profile:
    head:
      summary: WEB_AUTH discovery and profile freshness check
      description: |
        Used by OpenVPN Connect to discover WEB_AUTH support and to check whether
        a previously-issued profile is still valid without triggering re-authentication.

        **Discovery (no token)**: Returns `Ovpn-WebAuth: <SITE_NAME>,external` in the
        response, telling OpenVPN Connect it must authenticate via an external browser.

        **Freshness check (with token)**: If a valid `VPN-Session-Token` header is
        supplied and the associated certificate has not expired, the `Ovpn-WebAuth`
        header is omitted, indicating the profile is still fresh.

        **Service Separation**: Only available on the user service.
      tags:
        - OpenVPN Connect WEB_AUTH
      parameters:
        - in: header
          name: VPN-Session-Token
          schema:
            type: string
            format: uuid
          required: false
          description: |
            UUID token returned as `VPN-Session-Token` by `GET /download` after a
            successful profile download. OpenVPN Connect sends this on subsequent
            HEAD requests to check whether the certificate is still valid.
      responses:
        '200':
          description: |
            Always 200. If `Ovpn-WebAuth` header is present, re-authentication is
            required. If absent, the existing profile is still valid.
          headers:
            Ovpn-WebAuth:
              schema:
                type: string
                example: "My VPN,external"
              description: |
                Present when re-authentication is required.
                Format: `<SITE_NAME>,external`
    get:
      summary: WEB_AUTH profile provisioning (browser flow)
      description: |
        Invoked by OpenVPN Connect in the user's external browser after the HEAD
        discovery response indicated re-authentication was required.

        **Unauthenticated**: Device metadata (deviceID, deviceModel, deviceOS,
        appVersion) is stashed in the session and the user is redirected to the
        OIDC login flow. On successful authentication, `/auth/callback` returns
        here via the `next_url` session key.

        **Authenticated**: A one-time `DownloadToken` is created with the user's
        OIDC groups (for template selection), then the browser is redirected to
        `openvpn://import-profile/https://<host>/download?token=<uuid>`. The OS
        passes this URL to OpenVPN Connect, which fetches the profile.

        **Service Separation**: Only available on the user service.
      tags:
        - OpenVPN Connect WEB_AUTH
      parameters:
        - in: query
          name: deviceID
          schema:
            type: string
          required: false
          description: OpenVPN Connect device identifier (stored for audit)
        - in: query
          name: deviceModel
          schema:
            type: string
          required: false
          description: Device model string (stored for audit)
        - in: query
          name: deviceOS
          schema:
            type: string
          required: false
          description: Device operating system (stored for audit)
        - in: query
          name: appVersion
          schema:
            type: string
          required: false
          description: OpenVPN Connect application version (stored for audit)
      responses:
        '302':
          description: |
            Redirect to OIDC login (unauthenticated) or to
            `openvpn://import-profile/<download_url>` (authenticated).
          headers:
            Location:
              schema:
                type: string
              description: |
                `/auth/login` for unauthenticated requests, or
                `openvpn://import-profile/https://<host>/download?token=<uuid>`
                for authenticated requests.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PSK
      description: |
        Pre-shared key authentication for devices and servers.
        
        PSKs are generated through the administrative web interface and must be
        included in the Authorization header as a Bearer token:
        ```
        Authorization: Bearer your-psk-here
        ```
        
        Each PSK is associated with a specific hostname and can be revoked
        through the administrative interface.
...